Make `derive_parser` work with all optional arguments
The current implementation using `winnow`'s `opt` does not work for optional arguments that are in the middle of the command. For example,

```
bar{.cta}.red.op.pred   p, a{, b}, {!}c;
```

This is because `opt` is greedy, and will always match `{, b}` instead of `,{!} c`. This change switches to using `alt` in this situation.

The `alt` implementation will also be used for `[|q]` patterns, since those are marked as optional. I believe this will still work, even though not strictly necessary.